test(#14730): consolidate 3rd live duplicated test pair - legacy extract_titles into canons
38 legacy tests = 30 intent-duplicates (mapping in PR body) + 8 real edges
ported: html-comment block, numeric-start key, underscore and dashed YAML
key fixtures (the canonical tests did not test their own names), leading-
whitespace h1, h2/h3 never promoted, root-FM/slide-FM adjacency, slide
number 0.

Proof of non-loss: 38 + 34 + 15 = 87 before, 39 + 16 = 55 after.
